Adapt MultiBars tests to GroupedBars, StackedBars tests
